+++ ORBIT/vessels/tasks/__init__.py
@@ -11,4 +11,3 @@ they are related to.
 from ._exceptions import MissingComponent
 from .cable_laying import *
 from .scour_protection import *
-from .offshore_substation import *
